+++ src/migrations/index.ts
@@ -12,6 +12,7 @@ import * as migration_20260723_002255_add_research_queue_version from './2026072
 import * as migration_20260806_225622 from './20260806_225622';
 import * as migration_20260808_add_claims_vs_reality from './20260808_add_claims_vs_reality';
 import * as migration_20260809_162628 from './20260809_162628';
+import * as migration_20260809_182227 from './20260809_182227';
 
 export const migrations = [
   {
@@ -82,6 +83,11 @@ export const migrations = [
   {
     up: migration_20260809_162628.up,
     down: migration_20260809_162628.down,
-    name: '20260809_162628'
+    name: '20260809_162628',
+  },
+  {
+    up: migration_20260809_182227.up,
+    down: migration_20260809_182227.down,
+    name: '20260809_182227'
   },
 ];